# docs/_build/openapi.yaml
openapi: 3.0.0
info:
  title: DIXA API
  description: 'Project DIXA API, documentation management'
  version: 0.0.1
servers:
  - url: 'http://localhost:8000/api/v1'
    description: Servidor local
  - url: 'http://api.qa-dixa.puller.mx/api/v1'
    description: Servidor de qa
  - url: 'http://staging-dixa.puller.mx/api/v1'
    description: Servidor de staging
paths:
...
  /verify-auth:
    get:
      tags:
        - Auth
      summary: Verify Auth
      operationId: verifyAuth
      description: verify that the user is still authenticated
      responses:
        '200':
          description: Basic structure of response
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAuth:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    '400':
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: string
                example: Bad Request
              success:
                type: boolean
                example: false
    '401':
      description: unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: string
                example: Unauthorized
              success:
                type: boolean
                example: false
    '403':
      description: Not authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: string
                example: Forbidden
              success:
                type: boolean
                example: false
    '404':
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: string
                  example: Registro no encontrado
              success:
                type: boolean
                example: false
    '422':
      description: Error validation
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: string
                  example: El campo key_name es obligatorio
              success:
                type: boolean
                example: false
...
security:
  - bearerAuth: []